+++ src/lib.rs
@@ -161,4 +161,4 @@ pub mod tweet;
 pub mod user;
 
 pub use crate::auth::{Token, KeyPair};
-pub use crate::common::Response;
+pub use crate::common::{Response, RateLimit};