+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Various `request_*` and `response_*` functions to assemble and execute signed API requests
   - New type `ParamList` to represent a list of parameters to an API call
   - New `raw` example showing how to pull the raw JSON from fetching a tweet
+- `RateLimit` is now an exposed type with its own docs
 
 ### Changed
 - The repository has moved! egg-mode's new home is now at `egg-mode-rs/egg-mode`